Move RunnerProxyStore to RunnerBarCore (#1612)
- Add Sources/RunnerBarCore/Runner/RunnerProxyStore.swift (public access)
- Tombstone Sources/RunnerBar/Runner/RunnerProxyStore.swift
- Update docs/architecture/file-hierarchy.md

RunnerProxyStoreProtocol, RunnerProxyConfig, and RunnerProxyStoreError
are already in Core — this moves the concrete actor alongside them,
completing the proxy subsystem. No app-layer coupling: imports Foundation
and RunnerBarCore only. Access modifiers widened to public; singleton
pattern and all private helpers unchanged.
